Create Dockerfile
⌨️ Activity: Create a Dockerfile and open a pull request
Create a file titled action-a/Dockerfile by using this quick link or manually:

Create a new branch. Branches should be named intentionally, so a good name for this branch could be first-action.
On the new branch, create a directory: action-a. Note: If you're working on GitHub.com, you can create a directory and a file at the same time by naming the file action-a/Dockerfile.
In the action-a directory, create a file titled Dockerfile.